localStorage and much more
Removed unused functions & variables
Did some more styling
Changed image url since volumeinfo.imagesomething is not always available
Changed so that does not go to detail if focus is on input -> might change from on hold to on double click
Changed name
Sorted books by date

Currently 2 issues
When new book is added, it is sorted only if data is filled in. This is because release can come from google servers which are slower than the code causeing it to skip the sort function
When a book is added it will be added to local storage before it has the time to get the data, if above issue gets resolved, this will probably as well.