Major localStorage rework, details below
local storage was not working propperly due to it being a string instead of my objects. Fixed this by doing a foreach on the objects it contained

Comments.js class was in loop becasue was adding user with comments... Fixed this by creating a new user object in comments with same username & id, preventing a loop.

Added check on seed because of isbn conflict with store (duplicates but isbn is used as key & unique)

Due to having to create a completed object, had to add props in constructor of comments.js & bookposts.js

Added a few more places at which the object will be saved to the local store

Forgot to remove string check at store.js line 114, next commit is neglectable