Adjusted to feedback & small fix...
Adjusted BEM notation of item
Other issues mentioned in feedback (globalFunctions, context & onMouseDown) were already fixed in prior commit (week05 branch).

Forgot to change the updatebookdata from old bookPost update function to a new store update function. Also made it so that when updating the data you will also get the real title & data. did this because I personally prefer it this way.

App is also online @ https://migueldp.be/books

Note that https is required to copy the isbn numbers automaticly & it has been adjusted to how I want it to be as I will be the one using it. This means that I have removed the description & comments as I have no need for this (left the detail page but doubt ill use it). Online version should be save version as private branch on this repo.